extract snapshot methods into a separate mixin to be used in the new timeline view
