+++ frontend/src/pages/instance/VersionHistory/Snapshots/index.vue
@@ -48,18 +48,14 @@ import { mapState } from 'vuex'
 
 import InstanceApi from '../../../../api/instances.js'
 import SnapshotApi from '../../../../api/projectSnapshots.js'
-import SnapshotsApi from '../../../../api/snapshots.js'
 
 import EmptyState from '../../../../components/EmptyState.vue'
 import AssetCompareDialog from '../../../../components/dialogs/AssetCompareDialog.vue'
 import AssetDetailDialog from '../../../../components/dialogs/AssetDetailDialog.vue'
 import SnapshotEditDialog from '../../../../components/dialogs/SnapshotEditDialog.vue'
 import UserCell from '../../../../components/tables/cells/UserCell.vue'
-import { downloadData } from '../../../../composables/Download.js'
 import permissionsMixin from '../../../../mixins/Permissions.js'
-import Alerts from '../../../../services/alerts.js'
-import Dialog from '../../../../services/dialog.js'
-import Product from '../../../../services/product.js'
+import snapshotsMixin from '../../../../mixins/Snapshots.js'
 import { applySystemUserDetails } from '../../../../transformers/snapshots.transformer.js'
 import DaysSince from '../../../application/Snapshots/components/cells/DaysSince.vue'
 import DeviceCount from '../../../application/Snapshots/components/cells/DeviceCount.vue'
@@ -75,7 +71,7 @@ export default {
         AssetDetailDialog,
         AssetCompareDialog
     },
-    mixins: [permissionsMixin],
+    mixins: [permissionsMixin, snapshotsMixin],
     inheritAttrs: false,
     props: {
         instance: {
@@ -183,98 +179,6 @@ export default {
                 return acc
             }, {})
             return deviceCounts
-        },
-        // snapshot actions - delete
-        showDeleteSnapshotDialog (snapshot) {
-            Dialog.show({
-                header: 'Delete Snapshot',
-                text: 'Are you sure you want to delete this snapshot?',
-                kind: 'danger',
-                confirmLabel: 'Delete'
-            }, async () => {
-                await SnapshotsApi.deleteSnapshot(snapshot.id)
-                const index = this.snapshots.indexOf(snapshot)
-                this.snapshots.splice(index, 1)
-                Alerts.emit('Successfully deleted snapshot.', 'confirmation')
-            })
-        },
-        // snapshot actions - rollback
-        showRollbackDialog (snapshot) {
-            Dialog.show({
-                header: 'Restore Snapshot',
-                kind: 'danger',
-                text: `This will overwrite the current instance.
-                       All changes to the flows, settings and environment variables made since the last snapshot will be lost.
-                       Are you sure you want to deploy to this snapshot?`,
-                confirmLabel: 'Confirm'
-            }, async () => {
-                await SnapshotApi.rollbackSnapshot(this.instance.id, snapshot.id)
-                Alerts.emit('Successfully deployed snapshot.', 'confirmation')
-            })
-        },
-        // snapshot actions - set as device target
-        showDeviceTargetDialog (snapshot) {
-            Dialog.show({
-                header: 'Set Device Target Snapshot',
-                text: `Are you sure you want to set this snapshot as the device target?
-                       All devices assigned to this instance will be restarted on this snapshot.`,
-                confirmLabel: 'Set Target'
-            }, async () => {
-                await InstanceApi.updateInstanceDeviceSettings(this.instance.id, {
-                    targetSnapshot: snapshot.id
-                })
-                this.$emit('instance-updated')
-            })
-        },
-        async downloadSnapshotPackage (snapshot) {
-            Product.capture('ff-snapshot-download', {
-                'snapshot-id': snapshot.id
-            }, {
-                instance: this.instance?.id
-            })
-            const ss = await SnapshotsApi.getSummary(snapshot.id)
-            const owner = ss.device || ss.project
-            const ownerType = ss.device ? 'device' : 'instance'
-            const packageJSON = {
-                name: `${owner.safeName || owner.name}`.replace(/[^a-zA-Z0-9-]/g, '-').toLowerCase(),
-                description: `${ownerType} snapshot, ${snapshot.name} - ${snapshot.description}`,
-                private: true,
-                version: '0.0.0-' + snapshot.id,
-                dependencies: ss.modules || {}
-            }
-            downloadData(packageJSON, 'package.json')
-        },
-        showDownloadSnapshotDialog (snapshot) {
-            this.$refs.snapshotExportDialog.show(snapshot)
-        },
-        showViewSnapshotDialog (snapshot) {
-            Product.capture('ff-snapshot-view', {
-                'snapshot-id': snapshot.id
-            }, {
-                instance: this.instance?.id
-            })
-            SnapshotsApi.getFullSnapshot(snapshot.id).then((data) => {
-                this.$refs.snapshotViewerDialog.show(data)
-            }).catch(err => {
-                console.error(err)
-                Alerts.emit('Failed to get snapshot.', 'warning')
-            })
-        },
-        showCompareSnapshotDialog (snapshot) {
-            Product.capture('ff-snapshot-compare', {
-                'snapshot-id': snapshot.id
-            }, {
-                instance: this.instance?.id
-            })
-            SnapshotsApi.getFullSnapshot(snapshot.id)
-                .then((data) => this.$refs.snapshotCompareDialog.show(data, this.snapshotList))
-                .catch(err => {
-                    console.error(err)
-                    Alerts.emit('Failed to get snapshot.', 'warning')
-                })
-        },
-        showEditSnapshotDialog (snapshot) {
-            this.$refs.snapshotEditDialog.show(snapshot)
         },
         onSnapshotEdit (snapshot) {
             const index = this.snapshots.findIndex(s => s.id === snapshot.id)
